% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{qvs.values}
\alias{qvs.values}
\title{Generate finite set of mixture proportions}
\usage{
qvs.values(p, m)
}
\arguments{
\item{p}{number of populations, must be at least 2.}

\item{m}{number of different mixture proportions, must be at least 2.}
}
\value{
Returns a \code{p} by \code{m} matrix of mixture proportions.
}
\description{
Produces the finite set of mixture proportions for simulated data.
}
